packages/api/requirements.txt
python-dotenv==1.0.1
openai==1.26.0
fastapi==0.110.3
uvicorn==0.29.0
supabase==2.7.4